requirements.txt
ttkbootstrap==1.10.1
tkcalendar==1.6.1
pillow==10.0.0
openpyxl==3.1.2 